feat(api): 添加Blob类型支持及相关功能扩展
- 在ValueType枚举中添加Blob类型支持
- 为Blob类型实现hash、display、length等基础方法
- 在GetVerticesExecutor中添加Blob类型的空值判断
- 扩展C API添加按列索引获取结果和列类型的方法
- 添加数据库打开标志和配置选项
- 更新错误报告文档

+++ src/api/embedded/statistics.rs
@@ -186,7 +186,7 @@ impl QueryStatistics {
     pub fn from_metadata(metadata: &crate::api::core::ExecutionMetadata) -> Self {
         Self {
             rows_affected: metadata.rows_returned,
-            rows_returned: metadata.rows_returned as usize,
+            rows_returned: metadata.rows_returned as u64,
             ..Default::default()
         }
     }
